Extract common functionality between SCRAM-SHA-1 and SCRAM-SHA-256 into base class
